+++ app/Http/Controllers/Admin/AdminCompanyLocationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CompanyLocation;
+use App\Models\Company;
 
 class AdminCompanyLocationController extends Controller
 {
@@ -48,6 +49,10 @@ class AdminCompanyLocationController extends Controller
     }
 
     public function delete($id){
+        $count = Company::where('company_location_id',$id)->count();
+        if($count > 0){
+            return redirect()->back()->with('error','Invalid !! Companies exist with this location');
+        }
         CompanyLocation::where('id',$id)->delete();
         return redirect()->route('admin_company_location')->with('success','Data is deleted successfully');
     }
