+++ app/Http/Controllers/Admin/AdminJobTypeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\JobType;
+use App\Models\Job;
 
 class AdminJobTypeController extends Controller
 {
@@ -48,6 +49,10 @@ class AdminJobTypeController extends Controller
     }
 
     public function delete($id){
+        $count = Job::where('job_type_id',$id)->count();
+        if($count > 0){
+            return redirect()->back()->with('error','Invalid !! Jobs existed with this type');
+        }
         JobType::where('id',$id)->delete();
         return redirect()->route('admin_job_type')->with('success','Data is deleted successfully');
     }
